% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dfm.R
\name{create_dfm}
\alias{create_dfm}
\title{Creates a Document Feature Matrix}
\usage{
create_dfm(obj, num_ngrams = 5, remove_punct = T, ...)
}
\arguments{
\item{obj}{A \code{quanteda::corpus}-class object}

\item{num_ngrams}{Number of \code{quanteda::dfm}-class objects to create}
}
\value{
A \code{list} of \code{quanteda::dfm} -class objects
}
\description{
Creates a \code{list} of \code{quanteda::dfm} based on specified ngrams
}
