% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Module to extract text data from various file types}
\usage{
extract_text(.filepath = NULL, .pdfimage = F, .filetype = NULL)
}
\arguments{
\item{.filepath}{Character string containing the path to the file 
from which the text data is to be extracted}

\item{.pdfimage}{Boolean denoting if the file is an image, 
for use when \code{file_path == 'pdf'}}

\item{.filetype}{Character string to specify the file type}
}
\value{
text data
}
\description{
Takes in a file path and returns the text extracted from the document
}
\examples{
\dontrun{

oper655_readme <- "https://raw.githubusercontent.com/AFIT-R/oper655_fa2019/master/README.md"

Text = extract_text(.filepath = oper655_readme,
                    .filetype = "txt")

}

}
