% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_AWN.R
\name{gadget_AWN}
\alias{gadget_AWN}
\title{Text analysis app for the AFMC We Need Data}
\usage{
gadget_AWN(
  obj,
  rownames = TRUE,
  theme = "flatly",
  width = "100\%",
  height = "600px",
  css = NULL
)
}
\arguments{
\item{rownames}{\code{logical} Should rownames be included?}

\item{theme}{\code{character} A bootswatch theme provided to \code{shinythemes::shinytheme}}

\item{width}{\code{character} Width of the gadget (in valid css units)}

\item{height}{\code{character} Height of the gadget (in valid css units)}

\item{css}{\code{character} Path to a custom css file}

\item{data}{A data set}
}
\value{
A \code{list} of length 2
  \item{data}{A \code{data.frame} containing the columns that were not removed}
  \item{script}{A line of code that can be used to replicate cleaning performed in the gadget}

A printed shiny app
}
\description{
Shiny gadget used to visually inspect columns in a data set and select columns to remove
}
\examples{
\dontrun{clean_columns(mtcars)}

}
\concept{shinygadgets}
