% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfidf.R
\name{get_tfidf}
\alias{get_tfidf}
\title{Create a DFM with term frequency - inverse document frequency values}
\usage{
get_tfidf(dfm, docvar, ...)
}
\arguments{
\item{dfm}{A document feature matrix}

\item{docvar}{A character string denoting and document variable within \code{dfm}}

\item{...}{Not currently used}
}
\description{
Create a DFM with term frequency - inverse document frequency values
}
