% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topics.R
\name{get_topics}
\alias{get_topics}
\title{Perform topic modeling on a corpus}
\usage{
get_topics(corp, dfm_subset, num_topics, num_ngrams, ...)
}
\arguments{
\item{corp}{A corpus object}

\item{dfm_subset}{A logical expression indicating elements to keep: missing values are taken as false}

\item{num_topics}{The number of topics to return}

\item{num_ngrams}{The phrase length to}

\item{...}{Additional options passed to \code{Text.Replace::create_dfm()}}
}
\description{
Perform topic modeling on a corpus
}
\examples{
\dontrun{

 # load AFMC We Need data set
   csv = "path/to/AFMC_We_Need/data/file.csv"
 
 # Extract text from the CSV file  
   DATA = Text.Replace:::extract_text(csv)
 
 # Subset DATA for question 1 of the Field Survey as object \code{q1f_DATA}
   q1f_DATA = subset(DATA, Source == 'Field Survey' & Question == 1)

 # Create the corpus object
   corp = Text.Replace::create_corpus(q1f_DATA, "Comments")
   
 # Get the topics 
   Text.Replace::get_topics(corp = corp,
                            dfm_subset = Base == "Robins",
                            num_topics = 10,
                            num_ngrams = 1,
                            remove = quanteda::stopwords("english")) 

}
}
