% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases_collocations.R
\name{phrases_collocations}
\alias{phrases_collocations}
\title{Determines Collocated Terms}
\usage{
phrases_collocations(x, ngram = 2, min_count = 5)
}
\arguments{
\item{x}{A \code{quanteda::token}}

\item{ngram}{Ngram size to locate}

\item{min_count}{The minimum number of instances for each ngram to be included}
}
\value{
A \code{dataframe} containing located phrases and number of appearances.
}
\description{
Searches a \code{quanteda::tokens} to locate ngrams.
The function cleans the token and returns a \code{dataframe} with results.
}
