% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Computes present values}
\usage{
pv(FV, r, n = 5)
}
\arguments{
\item{FV}{A \code{numeric} vector of future values of the items}

\item{r}{A \code{numeric} vector (of length 1) for the rate}

\item{n}{The number of years that the rate \code{r} is applied to \code{FV}}
}
\description{
Computes present values given a future value, a rate and a number of years.
}
\examples{
\dontrun{
#Test for pv
pv(FV = 1000, r = .08, n = 5)

fv_l <- list(fv1 = 800, 
     fv2 = 900, 
     fv3 = 1100)

pv(fv_l, 0.08)
# Throws an error
}
}
