% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_phrases.R
\name{replace_phrases}
\alias{replace_phrases}
\title{Replace text values}
\usage{
replace_phrases(text, old, new)
}
\arguments{
\item{text}{}

\item{old}{}

\item{new}{}
}
\value{

}
\description{
Replace text values
}
