% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_sampler.R
\name{response_sampler}
\alias{response_sampler}
\title{Randomly sample survey responses}
\usage{
response_sampler(df, text_col, n)
}
\arguments{
\item{df}{A dataframe object with}

\item{text_col}{The column with each response per row}

\item{n}{The number of responses to sample}
}
\value{
A vector of sampled comments from the supplied dataframe.
}
\description{
This
}
\examples{
\dontrun{
data(austensurvey)
austen_2016 <- subset(austensurvey, subset = Year == 2016)
set.seed(100)
rand_2016 <- response_sampler(austen_2016, "Comment", 10)
}

}
