% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check_gadget.R
\name{spell_check_gadget}
\alias{spell_check_gadget}
\title{Shiny Spell Check Gadget}
\usage{
spell_check_gadget(words, corpus, my_dict = hunspell::dictionary("en_US"))
}
\arguments{
\item{words}{A \code{character} vector of strings to spell check}

\item{corpus}{A \code{quanteda corpus} to examine each word from words in 
its textual context.}

\item{my_dict}{A \code{hunspell} dictionary object to use as reference in
suggesting spelling corrections. The default is the 'en-US' dictionary.}
}
\value{
A dataframe
}
\description{
A \code{shiny gadget} that pairs with the
  \code{\link{spellcheck_words}} to examine flagged words and create a
  dataframe for string replacements.
}
