% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.Replace-Package.R
\docType{package}
\name{Text.Replace}
\alias{Text.Replace}
\alias{Text.Replace-package}
\title{Analyze Survey Data by replacing standard phrases}
\description{
This package imports and analyzes survey data with 
             associated metadata.
}
\details{
This section should provide a more detailed overview of how to use the
        package, including the most important functions.
}
\author{
Clayton Wilson
        Maintainer: Clayton Wilson
}
\keyword{package}
