% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_frequency.R
\name{view_frequency}
\alias{view_frequency}
\title{View the frequency of each document variable in the data set}
\usage{
view_frequency(dfm, docvar, ...)
}
\arguments{
\item{dfm}{A document feature matrix created by \code{create_dfm()}}

\item{docvar}{A character string denoting the document variable to be displayed}

\item{...}{Additional arguments sent to \code{metricsgraphics::mjs_plot()}}
}
\description{
View the frequency of each document variable in the data set
}
