% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_count}
\alias{word_count}
\title{Return a count of the number of words in each document}
\usage{
word_count(.data, .column = NULL)
}
\arguments{
\item{.data}{A character vector or a \code{data.frame} containing a character vector}

\item{.column}{Character string naming a column from \code{.data}
If \code{.column = NULL} \code{.data} is presumed to be a character vector}
}
\value{
A \code{data.frame} to which a column containing the number of words 
       for each document will be added to the column in \code{data}.  If 
       \code{.data} is a character vector the object returned will be a two-column
       \code{data.frame} in which the first column contains the text from each document
       and the second column will contain
}
\description{
Return a count of the number of words in each document
}
