% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corpus.R
\name{create_corpus}
\alias{create_corpus}
\title{Build a corpus object}
\usage{
create_corpus(.data, .column)
}
\arguments{
\item{.data}{A data frame containing header values and single comments column}

\item{.column}{A \code{string} value identifying column with text to analyze}
}
\value{
A \code{quanteda} corpus object
}
\description{
Creates a \code{quanteda::corpus}-class object from a collection of texts
}
